pyproject.toml
[tool.poetry]
name = "whisper-x-rep"
version = "0.1.0"
description = ""
authors = ["bonchardon <marynaboroda123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.74.0"
pydub = "^0.25.1"
loguru = "^0.7.3"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
